+++ migrations/1_ERC1155Inventory.js
@@ -1,0 +1,9 @@
+const ERC1155Inventory = artifacts.require('ERC1155InventoryMock');
+
+module.exports = async (deployer, network, accounts) => {
+    const [creator, minter] = accounts;
+
+    await deployer.deploy(ERC1155Inventory);
+    const inventory = await ERC1155Inventory.deployed();
+    await inventory.addMinter(minter, {from: creator});
+}