+++ test/contracts/token/ERC1155/implementations/old_ERC1155PausableInventory.js
@@ -1,3 +1,0 @@
-const implementation = require('./old_ERC1155Inventory');
-implementation.contract = "ERC1155PausableInventoryMock";
-module.exports = implementation;