+++ test/contracts/token/ERC1155721/implementations/old_ERC1155721PausableInventory.js
@@ -1,5 +1,0 @@
-const implementation = require('./old_ERC1155721Inventory');
-implementation.contract = "PausableInventoryMock";
-implementation.name = "PausableInventoryMock";
-implementation.symbol = "PIM";
-module.exports = implementation;