+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## 5.0.0
 
 ### Breaking changes
+ * Removed `ERC721`, `ERC1155AssetsInventory`, `AssetsInventory`, and their derived contracts.
+
+### New features
+ * Added an ERC721-compliant, NFTs only `_batchMint` function for `ERC1155721Inventory`.
+
+## 5.0.0
+
+### Breaking changes
  * Updated to `@animoca/ethereum-contracts-core_library@3.1.1`.
 
 ### New features